+++ src/components/Banner/index.ts
@@ -1,0 +1,2 @@
+export { default as Banner, type BannerRootProps, type BannerMediaProps } from './Banner';
+export { BannerVariants, BannerMediaSize } from './types';