+++ src/components/Banner/types.ts
@@ -1,0 +1,9 @@
+export enum BannerVariants {
+  inline = 'inline',
+  top = 'top',
+}
+
+export enum BannerMediaSize {
+  half = '1/2',
+  third = '1/3',
+}