+++ src/providers/SeldonProvider/SeldonProvider.test.tsx
@@ -1,15 +1,16 @@
 import { render, screen } from '@testing-library/react';
 import { SeldonProvider } from './SeldonProvider';
 import { SSRMediaQuery } from './utils';
+import { Breakpoints } from '../../utils';
 
 describe('SeldonProvider', () => {
   describe('Media Queries', () => {
     it('three children render with the fresnel classes', () => {
       render(
         <SeldonProvider>
-          <SSRMediaQuery.Media lessThan="md">Small Breakpoint</SSRMediaQuery.Media>
-          <SSRMediaQuery.Media between={['md', 'lg']}>Medium Breakpoint</SSRMediaQuery.Media>
-          <SSRMediaQuery.Media greaterThanOrEqual="lg">Large Breakpoint</SSRMediaQuery.Media>
+          <SSRMediaQuery.Media lessThan={Breakpoints.md}>Small Breakpoint</SSRMediaQuery.Media>
+          <SSRMediaQuery.Media between={[Breakpoints.md, Breakpoints.lg]}>Medium Breakpoint</SSRMediaQuery.Media>
+          <SSRMediaQuery.Media greaterThanOrEqual={Breakpoints.lg}>Large Breakpoint</SSRMediaQuery.Media>
         </SeldonProvider>,
       );